Added TimedInterrupt and renamed Automatic to AlternatingInterrupt
A timed interrupt basically just ignores an interrupt for a set amount of time. This might be useful if a static time might need to be waited to avoid activating twice on the same thing. Automatic was renamed to Alternating and given the functionality to actually "alternate" over a certain amount of edges before triggering again. The original automatic interrupt really did nothing as an opposite interrupt cannot physically be seen twice before the original one.